pyproject.toml
[tool.poetry]
name = "shadow-tech-test"
version = "0.0.1"
description = "Public Library Management API"
authors = ["Kaci Adjou <adjou.kaci@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
SQLAlchemy = "^2.0.28"
pydantic = "^2.6.4"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
pylint = "^3.1.0"
pytest = "^8.1.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
